pyproject.toml
[project]
name = "pnlcalib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "coloredlogs==15.0.1",
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "filelock==3.17.0",
    "flatbuffers==23.3.3",
    "fonttools==4.55.8",
    "fsspec==2025.2.0",
    "hpack==4.0.0",
    "humanfriendly==10.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.5",
    "kiwisolver==1.4.7",
    "lsq-ellipse==2.2.1",
    "markupsafe==3.0.2",
    "matplotlib==3.9.4",
    "mpmath==1.3.0",
    "munkres==1.1.4",
    "networkx==3.2.1",
    "numpy==2.0.2",
    "opencv-python==4.10.0.84",
    "packaging==24.2",
    "pillow==11.1.0",
    "pip>=25.0.1",
    "pyparsing==3.2.1",
    "pytest>=8.3.5",
    "python-dateutil==2.9.0.post0",
    "pyyaml==6.0.2",
    "ruff>=0.11.2",
    "scipy==1.13.1",
    "shapely==2.0.7",
    "six==1.17.0",
    "soccernet>=0.1.62",
    "sympy==1.13.3",
    "torch==2.3.1",
    "torchaudio==2.3.1",
    "torchvision==0.18.1",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "wandb>=0.19.8",
    "zipp==3.21.0",
    "zstandard==0.22.0",
]
